requirements.txt
altair==4.1.0
argcomplete==1.10.0
argon2-cffi==20.1.0
asgiref==3.4.1
astor==0.8.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.8.0
bleach==3.3.1
blinker==1.4
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
cycler==0.10.0
cymem==2.0.5
debugpy==1.3.0
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.15
docx2txt==0.8
entrypoints==0.3
extract-msg==0.23.1
gensim==4.0.1
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
IMAPClient==2.1.0
ipykernel==6.0.3
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.9.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==2.0.3
murmurhash==1.0.5
nbclient==0.5.3
nbconvert==6.5.1
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.6
notebook==6.4.12
numpy==1.22.0
olefile==0.46
packaging==21.0
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathy==0.6.0
pdfminer.six==20181108
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.1.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.18.3
ptyprocess==0.7.0
pyarrow==4.0.1
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.2
pydeck==0.6.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-pptx==0.6.18
pytz==2021.1
pyzmq==22.1.0
regex==2021.7.6
requests==2.26.0
scipy==1.7.0
seaborn==0.11.1
Send2Trash==1.7.1
six==1.12.0
smart-open==5.1.0
smmap==4.0.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.1.1
spacy-legacy==3.0.8
SpeechRecognition==3.8.1
sqlparse==0.4.2
srsly==2.4.1
streamlit==1.11.1
tenacity==8.0.1
terminado==0.10.1
testpath==0.5.0
textdistance==4.2.1
textract==1.6.3
thinc==8.0.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
tzlocal==1.5.1
urllib3==1.26.6
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
xlrd==1.2.0
XlsxWriter==1.4.4
